the script is patrol
So also, while using unity, it's not printing to the console. how do you
do that? and also, I realize there needs to be lots more code like
destination stuff instead of just moving along the x and y, that's easy
enough to put in once i have the boolean switching happening